--- Assets/Ipocom/Runtime/FrameMessage.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace Ipocom
-{
-    [Serializable]
-    public struct FrameMessage
-    {
-        public SonyMotionFormat.Head head;
-        public SonyMotionFormat.Sndf sndf;
-        public SonyMotionFormat.Fram fram;
-    }
-}
