+++ Assets/Ipocom/Runtime/SonyMotionFormat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 846c29fbb84eed34d8c51b3f98499fee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
