+++ Assets/Ipocom/Runtime/SonyMotionFormat/Bndt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8eb57223bd71ca409c9d0776795183e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
