+++ Assets/Ipocom/Runtime/SonyMotionFormat/Bnid.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed3bc46fd944842419d3a52de422949e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
