+++ Assets/Ipocom/Runtime/SonyMotionFormat/Box.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f01e2c60ba1dd744b49eddc0464805f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
