+++ Assets/Ipocom/Runtime/SonyMotionFormat/BoxTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 663ea065908d47049a4f513f4ba1620a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
