+++ Assets/Ipocom/Runtime/SonyMotionFormat/Btdt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a92fee6f8197f324483b48c16b5cf1ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
