+++ Assets/Ipocom/Runtime/SonyMotionFormat/Definition.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 09f6a10700211154f8d255d407302f17
+guid: d0cba9d35edede14c860bf27559d604b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
