+++ Assets/Ipocom/Runtime/SonyMotionFormat/Fram.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b412eff0a8052f48964dd522be8ec05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
