+++ Assets/Ipocom/Runtime/SonyMotionFormat/FrameMessage.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Ipocom.SonyMotionFormat
+{
+    [Serializable]
+    public struct FrameMessage
+    {
+        public Head head;
+        public Sndf sndf;
+        public Fram fram;
+    }
+}
