+++ Assets/Ipocom/Runtime/SonyMotionFormat/FrameMessage.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 389ef64088679584f8324553719f52f5
+guid: 46e11f08aa2d84745b0242702ceaba42
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
