+++ Assets/Ipocom/Runtime/SonyMotionFormat/Head.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acae757dc3cfe3f4aafc35c9e52af899
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
