+++ Assets/Ipocom/Runtime/SonyMotionFormat/Parser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4dbd75c8b3dee640a6e8207ca025f41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
