+++ Assets/Ipocom/Runtime/SonyMotionFormat/Pbid.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c180bae2f4f9c0d4a9d777fe432fc3d4
+guid: 421222bcfac94914184df389cdaf7a6f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
