+++ Assets/Ipocom/Runtime/SonyMotionFormat/Skdf.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40d2cd183e9e86e48be4cd8831e4482c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
