+++ Assets/Ipocom/Runtime/SonyMotionFormat/SkeletonMessage.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Ipocom.SonyMotionFormat
+{
+    [Serializable]
+    public struct SkeletonMessage
+    {
+        public Head head;
+        public Sndf sndf;
+        public Skdf skdf;
+    }
+}
