+++ Assets/Ipocom/Runtime/SonyMotionFormat/SkeletonMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfc62662a2a67384fad8b3b86e11f6b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
