+++ Assets/Ipocom/Runtime/SonyMotionFormat/Sndf.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e55b92ed7c1d3a48a4bbbf5315e9608
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
