+++ Assets/Ipocom/Runtime/SonyMotionFormat/Tran.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f26aa09dc8232ea4d9ef55fc52f892c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
